+++ github.com/lightstep/lightstep-tracer-go/lightstep/rand/num_gen.go
@@ -2,6 +2,6 @@ package rand
 
 // NumberGenerator defines an interface to generate numbers.
 type NumberGenerator interface {
-	Int63() int64
-	TwoInt63() (int64, int64)
+	Uint64() uint64
+	TwoUint64() (uint64, uint64)
 }