+++ github.com/lightstep/lightstep-tracer-go/proto_logencoder.go
@@ -2,8 +2,9 @@ package lightstep
 
 import (
 	"encoding/json"
+	"fmt"
 
-	cpb "github.com/lightstep/lightstep-tracer-go/collectorpb"
+	"github.com/lightstep/lightstep-tracer-common/golang/gogo/collectorpb"
 	"github.com/opentracing/opentracing-go/log"
 )
 
@@ -13,98 +14,142 @@ const (
 
 // An implementation of the log.Encoder interface
 type grpcLogFieldEncoder struct {
-	converter       *protoConverter
-	buffer          *reportBuffer
-	currentKeyValue *cpb.KeyValue
+	converter *protoConverter
+	buffer    *reportBuffer
+	keyValues []*collectorpb.KeyValue
 }
 
 func marshalFields(
 	converter *protoConverter,
-	protoLog *cpb.Log,
+	protoLog *collectorpb.Log,
 	fields []log.Field,
 	buffer *reportBuffer,
 ) {
 	logFieldEncoder := grpcLogFieldEncoder{
 		converter: converter,
 		buffer:    buffer,
+		keyValues: make([]*collectorpb.KeyValue, 0, len(fields)),
 	}
-	protoLog.Fields = make([]*cpb.KeyValue, len(fields))
-	for i, field := range fields {
-		logFieldEncoder.currentKeyValue = &cpb.KeyValue{}
+	for _, field := range fields {
 		field.Marshal(&logFieldEncoder)
-		protoLog.Fields[i] = logFieldEncoder.currentKeyValue
 	}
+	protoLog.Fields = logFieldEncoder.keyValues
 }
 
 func (lfe *grpcLogFieldEncoder) EmitString(key, value string) {
-	lfe.emitSafeKey(key)
-	lfe.emitSafeString(value)
+	var keyValue collectorpb.KeyValue
+	lfe.setSafeKey(&keyValue, key)
+	lfe.setSafeStringValue(&keyValue, value)
+	lfe.emitKeyValue(&keyValue)
 }
+
 func (lfe *grpcLogFieldEncoder) EmitBool(key string, value bool) {
-	lfe.emitSafeKey(key)
-	lfe.currentKeyValue.Value = &cpb.KeyValue_BoolValue{value}
+	var keyValue collectorpb.KeyValue
+	lfe.setSafeKey(&keyValue, key)
+	keyValue.Value = &collectorpb.KeyValue_BoolValue{BoolValue: value}
+	lfe.emitKeyValue(&keyValue)
 }
+
 func (lfe *grpcLogFieldEncoder) EmitInt(key string, value int) {
-	lfe.emitSafeKey(key)
-	lfe.currentKeyValue.Value = &cpb.KeyValue_IntValue{int64(value)}
+	var keyValue collectorpb.KeyValue
+	lfe.setSafeKey(&keyValue, key)
+	keyValue.Value = &collectorpb.KeyValue_IntValue{IntValue: int64(value)}
+	lfe.emitKeyValue(&keyValue)
 }
+
 func (lfe *grpcLogFieldEncoder) EmitInt32(key string, value int32) {
-	lfe.emitSafeKey(key)
-	lfe.currentKeyValue.Value = &cpb.KeyValue_IntValue{int64(value)}
+	var keyValue collectorpb.KeyValue
+	lfe.setSafeKey(&keyValue, key)
+	keyValue.Value = &collectorpb.KeyValue_IntValue{IntValue: int64(value)}
+	lfe.emitKeyValue(&keyValue)
 }
+
 func (lfe *grpcLogFieldEncoder) EmitInt64(key string, value int64) {
-	lfe.emitSafeKey(key)
-	lfe.currentKeyValue.Value = &cpb.KeyValue_IntValue{int64(value)}
+	var keyValue collectorpb.KeyValue
+	lfe.setSafeKey(&keyValue, key)
+	keyValue.Value = &collectorpb.KeyValue_IntValue{IntValue: value}
+	lfe.emitKeyValue(&keyValue)
 }
+
+// N.B. We are using a string encoding for 32- and 64-bit unsigned
+// integers because it will require a protocol change to treat this
+// properly. Revisit this after the OC/OT merger.  LS-1175
+//
+// We could safely continue using the int64 value to represent uint32
+// without breaking the stringified representation, but for
+// consistency with uint64, we're encoding all unsigned integers as
+// strings.
 func (lfe *grpcLogFieldEncoder) EmitUint32(key string, value uint32) {
-	lfe.emitSafeKey(key)
-	lfe.currentKeyValue.Value = &cpb.KeyValue_IntValue{int64(value)}
+	var keyValue collectorpb.KeyValue
+	lfe.setSafeKey(&keyValue, key)
+	keyValue.Value = &collectorpb.KeyValue_StringValue{StringValue: fmt.Sprint(value)}
+	lfe.emitKeyValue(&keyValue)
 }
+
 func (lfe *grpcLogFieldEncoder) EmitUint64(key string, value uint64) {
-	lfe.emitSafeKey(key)
-	lfe.currentKeyValue.Value = &cpb.KeyValue_IntValue{int64(value)}
+	var keyValue collectorpb.KeyValue
+	lfe.setSafeKey(&keyValue, key)
+	keyValue.Value = &collectorpb.KeyValue_StringValue{StringValue: fmt.Sprint(value)}
+	lfe.emitKeyValue(&keyValue)
 }
+
 func (lfe *grpcLogFieldEncoder) EmitFloat32(key string, value float32) {
-	lfe.emitSafeKey(key)
-	lfe.currentKeyValue.Value = &cpb.KeyValue_DoubleValue{float64(value)}
+	var keyValue collectorpb.KeyValue
+	lfe.setSafeKey(&keyValue, key)
+	keyValue.Value = &collectorpb.KeyValue_DoubleValue{DoubleValue: float64(value)}
+	lfe.emitKeyValue(&keyValue)
 }
+
 func (lfe *grpcLogFieldEncoder) EmitFloat64(key string, value float64) {
-	lfe.emitSafeKey(key)
-	lfe.currentKeyValue.Value = &cpb.KeyValue_DoubleValue{float64(value)}
+	var keyValue collectorpb.KeyValue
+	lfe.setSafeKey(&keyValue, key)
+	keyValue.Value = &collectorpb.KeyValue_DoubleValue{DoubleValue: value}
+	lfe.emitKeyValue(&keyValue)
 }
+
 func (lfe *grpcLogFieldEncoder) EmitObject(key string, value interface{}) {
-	lfe.emitSafeKey(key)
+	var keyValue collectorpb.KeyValue
+	lfe.setSafeKey(&keyValue, key)
 	jsonBytes, err := json.Marshal(value)
 	if err != nil {
 		emitEvent(newEventUnsupportedValue(key, value, err))
 		lfe.buffer.logEncoderErrorCount++
-		lfe.emitSafeString("<json.Marshal error>")
+		lfe.setSafeStringValue(&keyValue, "<json.Marshal error>")
+		lfe.emitKeyValue(&keyValue)
 		return
 	}
-	lfe.emitSafeJSON(string(jsonBytes))
+	lfe.setSafeJSONValue(&keyValue, string(jsonBytes))
+	lfe.emitKeyValue(&keyValue)
 }
 func (lfe *grpcLogFieldEncoder) EmitLazyLogger(value log.LazyLogger) {
 	// Delegate to `value` to do the late-bound encoding.
 	value(lfe)
 }
 
-func (lfe *grpcLogFieldEncoder) emitSafeKey(key string) {
-	if len(key) > lfe.converter.maxLogKeyLen {
-		key = key[:(lfe.converter.maxLogKeyLen-1)] + ellipsis
-	}
-	lfe.currentKeyValue.Key = key
-}
-func (lfe *grpcLogFieldEncoder) emitSafeString(str string) {
-	if len(str) > lfe.converter.maxLogValueLen {
+func (lfe *grpcLogFieldEncoder) setSafeStringValue(keyValue *collectorpb.KeyValue, str string) {
+	if lfe.converter.maxLogValueLen > 0 && len(str) > lfe.converter.maxLogValueLen {
 		str = str[:(lfe.converter.maxLogValueLen-1)] + ellipsis
 	}
-	lfe.currentKeyValue.Value = &cpb.KeyValue_StringValue{str}
+	keyValue.Value = &collectorpb.KeyValue_StringValue{StringValue: str}
 }
-func (lfe *grpcLogFieldEncoder) emitSafeJSON(json string) {
-	if len(json) > lfe.converter.maxLogValueLen {
+
+func (lfe *grpcLogFieldEncoder) setSafeJSONValue(keyValue *collectorpb.KeyValue, json string) {
+	if lfe.converter.maxLogValueLen > 0 && len(json) > lfe.converter.maxLogValueLen {
 		str := json[:(lfe.converter.maxLogValueLen-1)] + ellipsis
-		lfe.currentKeyValue.Value = &cpb.KeyValue_StringValue{str}
+		keyValue.Value = &collectorpb.KeyValue_StringValue{StringValue: str}
 		return
 	}
-	lfe.currentKeyValue.Value = &cpb.KeyValue_JsonValue{json}
+	keyValue.Value = &collectorpb.KeyValue_JsonValue{JsonValue: json}
+}
+
+func (lfe *grpcLogFieldEncoder) setSafeKey(keyValue *collectorpb.KeyValue, key string) {
+	if lfe.converter.maxLogKeyLen > 0 && len(key) > lfe.converter.maxLogKeyLen {
+		keyValue.Key = key[:(lfe.converter.maxLogKeyLen-1)] + ellipsis
+		return
+	}
+	keyValue.Key = key
+}
+
+func (lfe *grpcLogFieldEncoder) emitKeyValue(keyValue *collectorpb.KeyValue) {
+	lfe.keyValues = append(lfe.keyValues, keyValue)
 }