+++ github.com/lightstep/lightstep-tracer-go/raw_span.go
@@ -3,7 +3,7 @@ package lightstep
 import (
 	"time"
 
-	opentracing "github.com/opentracing/opentracing-go"
+	"github.com/opentracing/opentracing-go"
 )
 
 // RawSpan encapsulates all state associated with a (finished) LightStep Span.
@@ -38,8 +38,14 @@ type SpanContext struct {
 	// A probabilistically unique identifier for a [multi-span] trace.
 	TraceID uint64
 
+	// Most significant bits of a 128-bit TraceID
+	TraceIDUpper uint64
+
 	// A probabilistically unique identifier for a span.
 	SpanID uint64
+
+	// Propagates sampling decision
+	Sampled string
 
 	// The span's associated baggage.
 	Baggage map[string]string // initialized on first use
@@ -68,5 +74,6 @@ func (c SpanContext) WithBaggageItem(key, val string) SpanContext {
 		newBaggage[key] = val
 	}
 	// Use positional parameters so the compiler will help catch new fields.
-	return SpanContext{c.TraceID, c.SpanID, newBaggage}
+
+	return SpanContext{c.TraceID, c.TraceIDUpper, c.SpanID, c.Sampled, newBaggage}
 }