+++ github.com/lightstep/lightstep-tracer-go/span.go
@@ -4,7 +4,7 @@ import (
 	"sync"
 	"time"
 
-	ot "github.com/opentracing/opentracing-go"
+	"github.com/opentracing/opentracing-go"
 	"github.com/opentracing/opentracing-go/log"
 )
 
@@ -13,12 +13,13 @@ import (
 type spanImpl struct {
 	tracer     *tracerImpl
 	sync.Mutex // protects the fields below
+	finished   bool
 	raw        RawSpan
 	// The number of logs dropped because of MaxLogsPerSpan.
 	numDroppedLogs int
 }
 
-func newSpan(operationName string, tracer *tracerImpl, sso []ot.StartSpanOption) *spanImpl {
+func newSpan(operationName string, tracer *tracerImpl, sso []opentracing.StartSpanOption) *spanImpl {
 	opts := newStartSpanOptions(sso)
 
 	// Start time.
@@ -37,6 +38,7 @@ func newSpan(operationName string, tracer *tracerImpl, sso []ot.StartSpanOption)
 		sp.raw.Context.TraceID = opts.SetTraceID
 		sp.raw.Context.SpanID = opts.SetSpanID
 		sp.raw.ParentSpanID = opts.SetParentSpanID
+		sp.raw.Context.Sampled = opts.SetSampled
 	}
 
 	// Look for a parent in the list of References.
@@ -46,10 +48,14 @@ func newSpan(operationName string, tracer *tracerImpl, sso []ot.StartSpanOption)
 ReferencesLoop:
 	for _, ref := range opts.Options.References {
 		switch ref.Type {
-		case ot.ChildOfRef, ot.FollowsFromRef:
-			refCtx := ref.ReferencedContext.(SpanContext)
+		case opentracing.ChildOfRef, opentracing.FollowsFromRef:
+			refCtx, ok := ref.ReferencedContext.(SpanContext)
+			if !ok {
+				break ReferencesLoop
+			}
 			sp.raw.Context.TraceID = refCtx.TraceID
 			sp.raw.ParentSpanID = refCtx.SpanID
+			sp.raw.Context.Sampled = refCtx.Sampled
 
 			if l := len(refCtx.Baggage); l > 0 {
 				sp.raw.Context.Baggage = make(map[string]string, l)
@@ -74,22 +80,39 @@ ReferencesLoop:
 	sp.raw.Start = startTime
 	sp.raw.Duration = -1
 	sp.raw.Tags = opts.Options.Tags
+
+	if tracer.opts.MetaEventReportingEnabled && !sp.IsMeta() {
+		opentracing.StartSpan(LSMetaEvent_SpanStartOperation,
+			opentracing.Tag{Key: LSMetaEvent_MetaEventKey, Value: true},
+			opentracing.Tag{Key: LSMetaEvent_SpanIdKey, Value: sp.raw.Context.SpanID},
+			opentracing.Tag{Key: LSMetaEvent_TraceIdKey, Value: sp.raw.Context.TraceID}).
+			Finish()
+	}
 	return sp
 }
 
-func (s *spanImpl) SetOperationName(operationName string) ot.Span {
+func (s *spanImpl) SetOperationName(operationName string) opentracing.Span {
 	s.Lock()
 	defer s.Unlock()
+
+	if s.finished {
+		return s
+	}
+
 	s.raw.Operation = operationName
 	return s
 }
 
-func (s *spanImpl) SetTag(key string, value interface{}) ot.Span {
+func (s *spanImpl) SetTag(key string, value interface{}) opentracing.Span {
 	s.Lock()
 	defer s.Unlock()
 
+	if s.finished {
+		return s
+	}
+
 	if s.raw.Tags == nil {
-		s.raw.Tags = ot.Tags{}
+		s.raw.Tags = opentracing.Tags{}
 	}
 	s.raw.Tags[key] = value
 	return s
@@ -104,7 +127,7 @@ func (s *spanImpl) LogKV(keyValues ...interface{}) {
 	s.LogFields(fields...)
 }
 
-func (s *spanImpl) appendLog(lr ot.LogRecord) {
+func (s *spanImpl) appendLog(lr opentracing.LogRecord) {
 	maxLogs := s.tracer.opts.MaxLogsPerSpan
 	if maxLogs == 0 || len(s.raw.Logs) < maxLogs {
 		s.raw.Logs = append(s.raw.Logs, lr)
@@ -120,13 +143,15 @@ func (s *spanImpl) appendLog(lr ot.LogRecord) {
 }
 
 func (s *spanImpl) LogFields(fields ...log.Field) {
-	lr := ot.LogRecord{
-		Fields: fields,
-	}
 	s.Lock()
 	defer s.Unlock()
-	if s.tracer.opts.DropSpanLogs {
+
+	if s.finished || s.tracer.opts.DropSpanLogs {
 		return
+	}
+
+	lr := opentracing.LogRecord{
+		Fields: fields,
 	}
 	if lr.Timestamp.IsZero() {
 		lr.Timestamp = time.Now()
@@ -135,22 +160,23 @@ func (s *spanImpl) LogFields(fields ...log.Field) {
 }
 
 func (s *spanImpl) LogEvent(event string) {
-	s.Log(ot.LogData{
+	s.Log(opentracing.LogData{
 		Event: event,
 	})
 }
 
 func (s *spanImpl) LogEventWithPayload(event string, payload interface{}) {
-	s.Log(ot.LogData{
+	s.Log(opentracing.LogData{
 		Event:   event,
 		Payload: payload,
 	})
 }
 
-func (s *spanImpl) Log(ld ot.LogData) {
+func (s *spanImpl) Log(ld opentracing.LogData) {
 	s.Lock()
 	defer s.Unlock()
-	if s.tracer.opts.DropSpanLogs {
+
+	if s.finished || s.tracer.opts.DropSpanLogs {
 		return
 	}
 
@@ -162,12 +188,12 @@ func (s *spanImpl) Log(ld ot.LogData) {
 }
 
 func (s *spanImpl) Finish() {
-	s.FinishWithOptions(ot.FinishOptions{})
+	s.FinishWithOptions(opentracing.FinishOptions{})
 }
 
 // rotateLogBuffer rotates the records in the buffer: records 0 to pos-1 move at
 // the end (i.e. pos circular left shifts).
-func rotateLogBuffer(buf []ot.LogRecord, pos int) {
+func rotateLogBuffer(buf []opentracing.LogRecord, pos int) {
 	// This algorithm is described in:
 	//    http://www.cplusplus.com/reference/algorithm/rotate
 	for first, middle, next := 0, pos, pos; first != middle; {
@@ -182,21 +208,21 @@ func rotateLogBuffer(buf []ot.LogRecord, pos int) {
 	}
 }
 
-func (s *spanImpl) FinishWithOptions(opts ot.FinishOptions) {
+func (s *spanImpl) FinishWithOptions(opts opentracing.FinishOptions) {
+	s.Lock()
+	defer s.Unlock()
+
+	if s.finished {
+		return
+	}
+
+	s.finished = true
+
 	finishTime := opts.FinishTime
 	if finishTime.IsZero() {
 		finishTime = time.Now()
 	}
 	duration := finishTime.Sub(s.raw.Start)
-
-	s.Lock()
-	defer s.Unlock()
-
-	// If the duration is already set, this span has already been finished.
-	// Return so we don't double submit the span.
-	if s.raw.Duration >= 0 {
-		return
-	}
 
 	for _, lr := range opts.LogRecords {
 		s.appendLog(lr)
@@ -216,7 +242,7 @@ func (s *spanImpl) FinishWithOptions(opts ot.FinishOptions) {
 		// about the dropped logs. This means that we are effectively dropping one
 		// more "new" log.
 		numDropped := s.numDroppedLogs + 1
-		s.raw.Logs[numOld] = ot.LogRecord{
+		s.raw.Logs[numOld] = opentracing.LogRecord{
 			// Keep the timestamp of the last dropped event.
 			Timestamp: s.raw.Logs[numOld].Timestamp,
 			Fields: []log.Field{
@@ -230,20 +256,31 @@ func (s *spanImpl) FinishWithOptions(opts ot.FinishOptions) {
 	s.raw.Duration = duration
 
 	s.tracer.RecordSpan(s.raw)
+	if s.tracer.opts.MetaEventReportingEnabled && !s.IsMeta() {
+		opentracing.StartSpan(LSMetaEvent_SpanFinishOperation,
+			opentracing.Tag{Key: LSMetaEvent_MetaEventKey, Value: true},
+			opentracing.Tag{Key: LSMetaEvent_SpanIdKey, Value: s.raw.Context.SpanID},
+			opentracing.Tag{Key: LSMetaEvent_TraceIdKey, Value: s.raw.Context.TraceID}).
+			Finish()
+	}
 }
 
-func (s *spanImpl) Tracer() ot.Tracer {
+func (s *spanImpl) Tracer() opentracing.Tracer {
 	return s.tracer
 }
 
-func (s *spanImpl) Context() ot.SpanContext {
+func (s *spanImpl) Context() opentracing.SpanContext {
 	return s.raw.Context
 }
 
-func (s *spanImpl) SetBaggageItem(key, val string) ot.Span {
-
+func (s *spanImpl) SetBaggageItem(key, val string) opentracing.Span {
 	s.Lock()
 	defer s.Unlock()
+
+	if s.finished {
+		return s
+	}
+
 	s.raw.Context = s.raw.Context.WithBaggageItem(key, val)
 	return s
 }
@@ -260,4 +297,8 @@ func (s *spanImpl) Operation() string {
 
 func (s *spanImpl) Start() time.Time {
 	return s.raw.Start
+}
+
+func (s *spanImpl) IsMeta() bool {
+	return s.raw.Tags["lightstep.meta_event"] != nil
 }