+++ github.com/lightstep/lightstep-tracer-go/tracer_helpers.go
@@ -1,9 +1,9 @@
 package lightstep
 
 import (
-	"golang.org/x/net/context"
+	"context"
 
-	opentracing "github.com/opentracing/opentracing-go"
+	"github.com/opentracing/opentracing-go"
 )
 
 // Flush forces a synchronous Flush.
@@ -18,7 +18,7 @@ func Flush(ctx context.Context, tracer opentracing.Tracer) {
 	}
 }
 
-// CloseTracer synchronously flushes the tracer, then terminates it.
+// Close synchronously flushes the tracer, then terminates it.
 func Close(ctx context.Context, tracer opentracing.Tracer) {
 	switch lsTracer := tracer.(type) {
 	case Tracer:
@@ -42,6 +42,7 @@ func GetLightStepAccessToken(tracer opentracing.Tracer) (string, error) {
 	}
 }
 
+// FlushLightStepTracer flushes the tracer
 // DEPRECATED: use Flush instead.
 func FlushLightStepTracer(tracer opentracing.Tracer) error {
 	switch lsTracer := tracer.(type) {
@@ -55,6 +56,7 @@ func FlushLightStepTracer(tracer opentracing.Tracer) error {
 	}
 }
 
+// CloseTracer closes the tracer
 // DEPRECATED: use Close instead.
 func CloseTracer(tracer opentracing.Tracer) error {
 	switch lsTracer := tracer.(type) {