+++ github.com/lightstep/lightstep-tracer-go/CHANGELOG.md
@@ -1,6 +1,55 @@
 # Changelog
 
-## [Pending Release](https://github.com/lightstep/lightstep-tracer-go/compare/v0.15.6...HEAD)
+## [Pending Release](https://github.com/lightstep/lightstep-tracer-go/compare/v0.24.0...HEAD)
+
+## [v0.24.0](https://github.com/lightstep/lightstep-tracer-go/compare/v0.23.0...v0.24.0)
+* Fix issue where the `ot-tracer-sampled` header was propagated as an empty string if it was not explicitly set [#269]
+## [v0.23.0](https://github.com/lightstep/lightstep-tracer-go/compare/v0.22.0...v0.23.0)
+* Only emit spans dropped within a status interval in the status event report [#265](https://github.com/lightstep/lightstep-tracer-go/pull/265)
+* Generate uint64 trace IDs, instead of int64 [#267](https://github.com/lightstep/lightstep-tracer-go/pull/267)
+
+## [v0.22.0](https://github.com/lightstep/lightstep-tracer-go/compare/v0.21.0...v0.22.0)
+* Propagate the sampled flag, and use it to determine whether or not to report spans
+* Fix bug where trace IDs were not correctly truncated
+
+## [v0.21.0](https://github.com/lightstep/lightstep-tracer-go/compare/v0.20.0...v0.21.0)
+* Allow users to disable metrics via environment variable: `LS_METRICS_ENABLED=false`
+
+## [v0.20.0](https://github.com/lightstep/lightstep-tracer-go/compare/v0.19.0...v0.20.0)
+* Adding support for reporting metrics to Lightstep
+* Updating opencensus dependency
+
+## [v0.19.0](https://github.com/lightstep/lightstep-tracer-go/compare/v0.18.1...v0.19.0)
+* Add flush duration to status report struct.
+* Add option to set `MaxBufferedSpans`.
+* Add the option to add default tags.
+
+## [v0.18.1](https://github.com/lightstep/lightstep-tracer-go/compare/v0.18.0...v0.18.1)
+* Adding support for configuring custom propagators
+
+## [v0.18.0](https://github.com/lightstep/lightstep-tracer-go/compare/v0.17.1...v0.18.0)
+* Adding support for B3 headers [#224](https://github.com/lightstep/lightstep-tracer-go/issues/224)
+* Fix OpenCensus to LighStep ID conversion
+* Added new constructor `CreateTracer` to propagate errors [#226](https://github.com/lightstep/lightstep-tracer-go/issues/226)
+
+## [v0.17.1](https://github.com/lightstep/lightstep-tracer-go/compare/v0.17.0...v0.17.1)
+* Fixes [#219](https://github.com/lightstep/lightstep-tracer-go/issues/219) so that there is no longer a data race when reading and writing from spans
+
+## [v0.17.0](https://github.com/lightstep/lightstep-tracer-go/compare/v0.16.0...v0.17.0)
+* Migrate dependency management from dep to gomod
+* Lazy loggers can emit 0 -> N log entries instead of just one.
+* Refactor Collector Client to allow for user-defined transports.
+
+## [v0.16.0](https://github.com/lightstep/lightstep-tracer-go/compare/v0.15.6...v0.16.0)
+* Thrift transport is eliminated
+* If no transport is specified, UseHTTP is now the default (was UseGRPC)
+* Requires go >= 1.7
+* Update to gogo-1.2.1, no longer using Google protobuf
+* Imports `context` via the standard library instead of `golang.org/x/net/context`
+* Fixes [#182](https://github.com/lightstep/lightstep-tracer-go/issues/182), so that `StartSpan` can now take `SpanReference`s to non-LightStep `SpanContext`s use
+* Adds experimental OpenCensus exporter
+* Access-Tokens are passed as headers to support improved load balancing
+* Support for custom TLS certificates.
 
 ## [v0.15.6](https://github.com/lightstep/lightstep-tracer-go/compare/v0.15.5...v0.15.6)